\name{g2chvcdata-package}
\alias{g2chvcdata-package}
\alias{g2chvcdata}
\docType{package}
\title{
Provide HDF5 data files for the hippocampus vs cortex project.
}
\description{
HDF5 files were converted from raw files (stored in separate repo).
}
\details{
\tabular{ll}{
Package: \tab g2chvcdata\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-15\cr
License: \tab What license is it under?\cr
}
}
\author{
Stephen Eglen
}
\references{
Forthcomng manuscript.
}
\keyword{ package }

